+++ alembic.ini
@@ -1,0 +1,58 @@
+# Alembic configuration — single consolidated setup (2026-07).
+# All migrations live in alembic/versions/; the DB URL is injected by
+# alembic/env.py from upstage_backend.global_config.DATABASE_URL.
+
+[alembic]
+# path to migration scripts (anchor to this file so cwd does not matter)
+script_location = %(here)s/alembic
+
+# sys.path path prepended for env.py (src layout)
+prepend_sys_path = %(here)s/src
+
+# single version location for all future revisions
+version_locations = %(here)s/alembic/versions
+
+# version path separator; "os" splits version_locations on os.pathsep
+version_path_separator = os
+
+# sqlalchemy.url intentionally absent — env.py overrides it at runtime
+
+[post_write_hooks]
+# post_write_hooks defines scripts or Python functions that are run
+# on newly generated revision scripts.  See the documentation for further
+# detail and examples
+
+# Logging configuration
+[loggers]
+keys = root,sqlalchemy,alembic
+
+[handlers]
+keys = console
+
+[formatters]
+keys = generic
+
+[logger_root]
+level = WARN
+handlers = console
+qualname =
+
+[logger_sqlalchemy]
+level = WARN
+handlers =
+qualname = sqlalchemy.engine
+
+[logger_alembic]
+level = INFO
+handlers =
+qualname = alembic
+
+[handler_console]
+class = StreamHandler
+args = (sys.stderr,)
+level = NOTSET
+formatter = generic
+
+[formatter_generic]
+format = %(levelname)-5.5s [%(name)s] %(message)s
+datefmt = %H:%M:%S
