+++ pyproject.toml
@@ -85,7 +85,6 @@ include = ["upstage_backend*"]
 line-length = 100
 target-version = "py312"
 extend-exclude = [
-    "alembic/versions",
     "migration_scripts",
     "build",
     "dist",
