--- src/upstage_backend/assets/db_migrations/228da801fb0c_create_asset_type_table.py
@@ -1,39 +1,0 @@
-"""Create Asset Type Table
-
-Revision ID: 228da801fb0c
-Revises: c37d2e3ebe8f
-Create Date: 2024-09-06 15:22:42.434487
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = "228da801fb0c"
-down_revision: Union[str, None] = "c37d2e3ebe8f"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "asset_type",
-        sa.Column("id", sa.BigInteger(), nullable=False, autoincrement=True),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), default=None),
-        sa.Column("file_location", sa.Text(), nullable=False),
-        sa.Column(
-            "created_on",
-            sa.TIMESTAMP(),
-            server_default=sa.text("(now() at time zone 'utc')"),
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-
-
-def downgrade() -> None:
-    op.drop_table("asset_type")
